Move swf.
