+++ flowy/__init__.py
@@ -1,9 +1,9 @@
 from flowy.backend.local import LocalWorkflow
-from flowy.backend.swf.config import SWFActivityConfig
-from flowy.backend.swf.config import SWFWorkflowConfig
-from flowy.backend.swf.starter import SWFWorkflowStarter
-from flowy.backend.swf.worker import SWFActivityWorker
-from flowy.backend.swf.worker import SWFWorkflowWorker
+from flowy.swf.config import SWFActivityConfig
+from flowy.swf.config import SWFWorkflowConfig
+from flowy.swf.starter import SWFWorkflowStarter
+from flowy.swf.worker import SWFActivityWorker
+from flowy.swf.worker import SWFWorkflowWorker
 from flowy.operations import finish_order
 from flowy.operations import first
 from flowy.operations import parallel_reduce
